Split midi2tracktionedit from augene-project and form its own project.
The problem was that Compose for Desktop project structuring is not quite
mature, and when there is another MPP module in the project it fails to
skip "metadata" part in the Compose for Desktop project, which results
in a bunch of "missing package java.*" kind of errors as well as
package resolution failures (it happens with okio here).

It can be "fixed" by kicking out all non-Compose projects from the same
project. To achieve that, the module reference to midi2tracktionedit
has to be moved out as package ID (and retrieve from `mavenLocal()`) too.