Implement "Hot Reload" on augene-player.
This should fix issue #2, or at least vastly improve.

With an example edit (that I cannot publish due to copyright) which loads
in 1 sec. for 7 OPNplug instances and 1 ayumi-juce instance, it reloads
in 168ms.

I should probably try much bigger songs like 10-ish sfizz instances, but
this in general looks promising.

The diff detection is quite immature and unsafe, so we would probably
introduce a switch on the GUI, whether we hot reload or not.